+++ src/XRP/model/send-xrp-details.ts
@@ -1,6 +1,6 @@
 import { BigInteger } from 'big-integer'
 import { Wallet } from 'xpring-common-js'
-import XrpMemo, { XRPMemo } from './xrp-memo'
+import XrpMemo from './xrp-memo'
 
 /**
  * Describes the fine grained details for sending money over the XRP ledger. The
@@ -11,11 +11,6 @@ export default interface SendXrpDetails {
   amount: BigInteger | number | string
   destination: string
   sender: Wallet
-
-  /**
-   * @deprecated Use `memosList` instead.
-   */
-  memos?: Array<XRPMemo>
 
   memoList?: Array<XrpMemo>
 }