+++ src/XRP/model/xrp-check-create.ts
@@ -1,7 +1,7 @@
 import { Utils } from 'xpring-common-js'
 import { CheckCreate } from '../Generated/web/org/xrpl/rpc/v1/transaction_pb'
-import { XRPCurrencyAmount } from '.'
-import XRPLNetwork from '../../Common/xrpl-network'
+import XrpCurrencyAmount from './xrp-currency-amount'
+import XrplNetwork from '../../Common/xrpl-network'
 
 /*
  * Represents a CheckCreate transaction on the XRP Ledger.
@@ -21,7 +21,7 @@ export default class XrpCheckCreate {
    */
   public static from(
     checkCreate: CheckCreate,
-    xrplNetwork: XRPLNetwork,
+    xrplNetwork: XrplNetwork,
   ): XrpCheckCreate | undefined {
     const destination = checkCreate.getDestination()?.getValue()?.getAddress()
     if (!destination) {
@@ -32,7 +32,7 @@ export default class XrpCheckCreate {
     const destinationXAddress = Utils.encodeXAddress(
       destination,
       destinationTag,
-      xrplNetwork == XRPLNetwork.Test || xrplNetwork == XRPLNetwork.Dev,
+      xrplNetwork == XrplNetwork.Test || xrplNetwork == XrplNetwork.Dev,
     )
     if (!destinationXAddress) {
       return undefined
@@ -42,7 +42,7 @@ export default class XrpCheckCreate {
     if (!sendMaxCurrencyAmount) {
       return undefined
     }
-    const sendMax = XRPCurrencyAmount.from(sendMaxCurrencyAmount)
+    const sendMax = XrpCurrencyAmount.from(sendMaxCurrencyAmount)
     if (!sendMax) {
       return undefined
     }
@@ -67,7 +67,7 @@ export default class XrpCheckCreate {
    */
   private constructor(
     readonly destinationXAddress: string,
-    readonly sendMax: XRPCurrencyAmount,
+    readonly sendMax: XrpCurrencyAmount,
     readonly expiration?: number,
     readonly invoiceId?: string,
   ) {}