+++ test/Xpring/xpring-client.test.ts
@@ -1,9 +1,9 @@
 import bigInt from 'big-integer'
 import { assert } from 'chai'
 import { FakeWallet } from 'xpring-common-js'
-import FakeXRPClient from '../XRP/fakes/fake-xrp-client'
+import FakeXrpClient from '../XRP/fakes/fake-xrp-client'
 import FakeXrpPayIdClient from '../PayID/fakes/fake-xrp-pay-id-client'
-import { testXRPTransaction } from '../XRP/fakes/fake-xrp-protobufs'
+import { testXrpTransaction } from '../XRP/fakes/fake-xrp-protobufs'
 import TransactionStatus from '../../src/XRP/transaction-status'
 import XpringClient from '../../src/Xpring/xpring-client'
 import RawTransactionStatus from '../../src/XRP/raw-transaction-status'
@@ -27,7 +27,7 @@ const fakeRawTransactionStatus = new RawTransactionStatus(
   true,
 )
 const fakePaymentHistoryValue = []
-const fakeGetPaymentValue = testXRPTransaction
+const fakeGetPaymentValue = testXrpTransaction
 /* eslint-enable @typescript-eslint/no-magic-numbers, @typescript-eslint/naming-convention */
 
 // An amount to send
@@ -45,9 +45,9 @@ const xrpError = new Error('Error from XRP')
 
 describe('Xpring Client', function (): void {
   it('send - success', async function (): Promise<void> {
-    // GIVEN a XpringClient composed of a fake PayIDClient and a fake XRPClient which will both succeed.
+    // GIVEN a XpringClient composed of a fake PayIDClient and a fake XrpClient which will both succeed.
     const expectedTransactionHash = fakeTransactionHash
-    const xrpClient = new FakeXRPClient(
+    const xrpClient = new FakeXrpClient(
       fakeBalance,
       fakePaymentStatus,
       expectedTransactionHash,
@@ -73,7 +73,7 @@ describe('Xpring Client', function (): void {
   it('send - failure in PayID', function (done): void {
     // GIVEN a XpringClient composed of a PayIDClient which will throw an error.
     const expectedTransactionHash = 'deadbeefdeadbeefdeadbeef'
-    const xrpClient = new FakeXRPClient(
+    const xrpClient = new FakeXrpClient(
       fakeBalance,
       fakePaymentStatus,
       expectedTransactionHash,
@@ -97,8 +97,8 @@ describe('Xpring Client', function (): void {
   })
 
   it('send - failure in XRP', function (done): void {
-    // GIVEN a XpringClient composed of an XRPClient which will throw an error.
-    const xrpClient = new FakeXRPClient(
+    // GIVEN a XpringClient composed of an XrpClient which will throw an error.
+    const xrpClient = new FakeXrpClient(
       fakeBalance,
       fakePaymentStatus,
       xrpError,
@@ -123,8 +123,8 @@ describe('Xpring Client', function (): void {
   })
 
   it('send - failure in both', function (done): void {
-    // GIVEN a XpringClient composed of an XRPClient and a PayID client which both throw errors.
-    const xrpClient = new FakeXRPClient(
+    // GIVEN a XpringClient composed of an XrpClient and a PayID client which both throw errors.
+    const xrpClient = new FakeXrpClient(
       fakeBalance,
       fakePaymentStatus,
       xrpError,
@@ -148,8 +148,8 @@ describe('Xpring Client', function (): void {
   })
 
   it('Constructor - XpringError thrown for mismatched networks', function (): void {
-    // GIVEN a PayIDClient and an XRPClient on different networks.
-    const xrpClient = new FakeXRPClient(
+    // GIVEN a PayIDClient and an XrpClient on different networks.
+    const xrpClient = new FakeXrpClient(
       fakeBalance,
       fakePaymentStatus,
       fakeTransactionHash,