+++ CHANGELOG.md
@@ -14,6 +14,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `XRPPayIDClientInterface` was deprecated for two releases and has been removed. Use `XrpPayIdClientInterface` instead.
 - `XRPPayIDClient` was deprecated for two releases and has been removed.  Use `XrpPayIdClient` instead.
 - `XRPLNetwork` was deprecated for two releases and has been removed.  Use `XrplNetwork` instead.
+- `memos` on `SendXrpDetails` was deprecated for two releases and has been removed. Use `memoList` instead.
+- `XRPAccountSet` was deprecated for two releases and has been removed. Use `XrpAccountSet` instead.
+- `XRPCurrencyAmount` was deprecated for two releases and has been removed. Use `XrpCurrencyAmount` instead.
+- `XRPCurrency` was deprecated for two releases and has been removed. Use `XrpCurrency` instead.
+- `XRPIssuedCurrency` was deprecated for two releases and has been removed. Use `XrpIssuedCurrency` instead.
+- `XRPMemo` was deprecated for two releases and has been removed. Use `XrpMemo` instead.
+- `XRPPathElement` was deprecated for two releases and has been removed. Use `XrpPathElement` instead.
+- `XRPPath` was deprecated for two releases and has been removed. Use `XrpPath` instead.
+- `XRPPayment` was deprecated for two releases and has been removed. Use `XrpPayment` instead.
+- `XRPSigner` was deprecated for two releases and has been removed. Use `XrpSigner` instead.
+- `XRPTransactionType` was deprecated for two releases and has been removed. Use `XrpTransactionType` instead.
+- `XRPTransaction` was deprecated for two releases and has been removed. Use `XrpTransaction` instead.
+- `XRPUtils` was deprecated for two releases and has been removed. Use `XrpUtils` instead.
+- `XRPClient` was deprecated for two releases and has been removed. Use `XrpClient` instead.
+- `XRPError` was deprecated for two releases and has been removed. Use `XrpError` instead.
+- `XRPErrorType` was deprecated for two releases and has been removed. Use `XrpErrorType` instead.
+- `XrpTransaction.account` and `XrpTransaction.sourceTag` were deprecated for two releases and have been removed.
+  Use the X-address encoded field `sourceXAddress` instead.
+- `XrpPayment.destination` and `XrpPayment.destinationTag` were deprecated for two releases and have been removed.
+  Use the X-address encoded field `destinationXAddress` instead.
+
 ## 4.3.2 - 2020-06-17
 
 ## Added